requirements.txt
# Requires Python 3.8.12
tensorflow==2.13.0
opencv-python==4.10.0.84
numpy==1.24.3
matplotlib==3.7.5
captcha==0.6.0
pillow==10.4.0
